+++ src/chains/ethereum/ethereum.config.ts
@@ -16,14 +16,14 @@ export namespace EthereumConfig {
   export const config: Config = {
     mainnet: {
       chainId: 1,
-      rpcUrl: `https://mainnet.infura.io/v3/${process.env.INFURA_KEY}`,
+      rpcUrl: `https://mainnet.infura.io/v3/`,
       tokenListType: 'URL',
       tokenListSource:
         'https://wispy-bird-88a7.uniswap.workers.dev/?url=http://tokens.1inch.eth.link',
     },
     kovan: {
       chainId: 42,
-      rpcUrl: `https://kovan.infura.io/v3/${process.env.INFURA_KEY}`,
+      rpcUrl: `https://kovan.infura.io/v3/`,
       tokenListType: 'FILE',
       tokenListSource: 'src/chains/ethereum/erc20_tokens_kovan.json',
     },